fix DropdownMenu `menu` type in generated types
When the component is generated, a class with the same name as the file
is created. This happened to match the name of the imported type
`DropdownMenu` causing the class to be used instead of the imported type
for the `menu` property. This caused type mismatches when trying to use
this property.